+++ src/BarChart.ts
@@ -1,5 +1,6 @@
 import * as d3 from 'd3';
 import {IClassAffiliation} from './DataStructures';
+import {createTooltip} from './utils';
 
 export class BarChart {
   private readonly $node: d3.Selection<any>;
@@ -32,41 +33,13 @@ export class BarChart {
       .attr('class', 'bar');
 
     const barColors = d3.scale.category10();
-    const tooltip = this.createTooltip();
+    createTooltip(this.$node, $bars, (d) => d.label + ': ' + d.count);
 
     $bars.attr('x', (d, i) => { return x(d.label); })
     .attr('y', (d) => { return y(d.count); })
     .attr('width', + x.rangeBand())
     .attr('height', (d) => { return this.height - y(d.count); })
-    .attr('fill', (d, i) => barColors(String(i)))
-    .on('mouseover', function() { tooltip.style('display', null); })
-    .on('mouseout', function() { tooltip.style('display', 'none'); })
-    .on('mousemove', function(d) {
-      const xPosition = d3.mouse(this)[0] - 15;
-      const yPosition = d3.mouse(this)[1] - 25;
-      tooltip.attr('transform', 'translate(' + xPosition + ',' + yPosition + ')');
-      tooltip.select('text').text(d.label + ': ' + d.count);
-    });
+    .attr('fill', (d, i) => barColors(String(i)));
     $bars.exit().remove();
-  }
-
-  createTooltip() {
-    // Prep the tooltip bits, initial display is hidden
-    const tooltip = this.$node.append('g')
-      .classed('bar-tooltip', true)
-      .style('display', 'none');
-
-    tooltip.append('rect')
-      .attr('width', 30)
-      .attr('height', 20)
-      .attr('fill', 'white');
-
-    tooltip.append('text')
-      .attr('x', 15)
-      .attr('dy', '1.2em')
-      .style('text-anchor', 'middle')
-      .attr('font-size', '12px')
-      .attr('font-weight', 'bold');
-    return tooltip;
   }
 }
