+++ src/styles/_confusionmatrix.scss
@@ -103,7 +103,6 @@
 
 .matrix-wrapper {
   grid-area: matrix;
-  //background: #82BF6E;
 }
 
 .matrix {
@@ -146,4 +145,3 @@
     width: 100%;
   }
 }
-
